Updated inline documentation (#28)
* removed unnecessary usage of _response object

* removed extra documentation from sample code

* added tests for getElementTitle & getElementAttribute

* removed extra whitespace

* updated documentation to pass dartdoc

* formatted dart files
