+++ pubspec.yaml
@@ -7,8 +7,8 @@ environment:
   sdk: ">=2.10.2 <3.0.0"
 
 dependencies:
-  html: ^0.14.0+3
-  http: ^0.12.0+4
+  html: ^0.14.0+4
+  http: ^0.12.2
 
 dev_dependencies:
   test: ^1.15.0
